got newmessage working
